remade debug menu so it's better
fixes a plethora of issues with the debug console.
