+++ resources/theme/master_screen_scale.json
@@ -546,7 +546,13 @@
         },
         "misc": {"border_width": "0", "shadow_width": "0"}
     },
-    "#debug_console": {"prototype": "#window_base_theme"},
+    "#debug_console": {
+        "prototype": "#window_base_theme", 
+        "misc": {
+            "enable_title_bar": "1",
+            "enable_close_button": "0"
+        }
+    },
     "#loading_window": {
         "colours": {"normal_border": "#00000000", "dark_bg": "#00000000"},
         "misc": {"border_width": "0", "shadow_width": "0", "enable_title_bar": "0"}
