+++ scripts/debug_commands/__init__.py
@@ -2,7 +2,7 @@ from typing import List
 
 from scripts.debug_commands.cat import CatsCommand
 from scripts.debug_commands.command import Command
-from scripts.debug_commands.eval import EvalCommand
+from scripts.debug_commands.eval import EvalCommand, UnderstandRisksCommand
 from scripts.debug_commands.fps import FpsCommand
 from scripts.debug_commands.help import HelpCommand
 from scripts.debug_commands.settings import ToggleCommand, SetCommand, GetCommand
