+++ scripts/game_structure/screen_settings.py
@@ -137,7 +137,7 @@ def set_display_mode(
             AllScreens.rebuild_all_screens()
 
             scripts.screens.screens_core.screens_core.rebuild_core()
-            scripts.debug_menu.debugmode.rebuild_console()
+            scripts.debug_console.debug_mode.rebuild_console()
 
             screen_name = source_screen.name.replace(" ", "_")
             new_screen: "Screens" = getattr(AllScreens, screen_name)
